Simplify I/O bridge, add UCFM binary format, optimize AVR emulation
- Replace over-engineered pipe/endpoint/transport/adapter framework with
  flat 8-byte mapping entries (MCU periph → host resource)
- Add io_mcu_ops_t vtable to eliminate all arch #ifdef from platform code
- Add UCFM binary firmware format with address range descriptors
- Auto-detect HEX vs BIN on load via fw_load()/fw_load_bytes()
- Default bridge config (UART0 console + GPIO2 LED) loads without web UI
- ESP32 UART0 bridge uses stdio (putchar) for console output
- Batch AVR peripheral ticks every 32 cycles (was every instruction)
- Compute SREG flags in bulk instead of per-bit set_flag() calls
- Place AVR hot path in IRAM on ESP32
- Split emu_task: tight loop without GDB overhead when no client connected
- Fix ESP32 watchdog: vTaskDelay(1) replaces taskYIELD()
- Fix UART double-send: callback handles GPIO only, poll handles UART
- Add web API test script (tests/test_webapi.sh)
- Update AVR examples to use _delay_ms, uart_hello prints 10 lines
- PC: 183 MHz emulated (11x real ATmega328P)
- ESP32: 0.9 MHz emulated (1.8x improvement)

+++ src/avr/avr_twi.c
@@ -360,7 +360,7 @@ avr_twi_t *avr_twi_init(avr_cpu_t *cpu, const avr_twi_config_t *config)
     return twi;
 }
 
-void avr_twi_tick(avr_cpu_t *cpu, avr_twi_t *twi, uint8_t cycles)
+void avr_twi_tick(avr_cpu_t *cpu, avr_twi_t *twi, uint16_t cycles)
 {
     if (twi->pending_op == TWI_OP_NONE || twi->cycles_remaining == 0)
         return;
